Reset Grid Push/Pull
Currently the reset push and pull classes are set with left or right
properties and value "0%". This causes a problem with complex grids
which uses multiple push and pull on md, lg, xl, ...
A fix for this is to set the left or right properties with the value
"auto" instead of "0%".