Updates to PersonaCard. Bug #371 plus:
Order of actions <li> and detail sections in the HTML are now irrelevant. Action has an attribute 'data-datailsTargetId' which must match the id of the target <ul>
Whichever action has the 'is-active' attribute will be set to visible on initial load
JS no longer tied to specific class names in CSS for which action/details to show (no need to change CSS if changing actions)
Made sure it works with multiple cards on a page (so long as IDs are unique)
Cleaned up CSS, JS and HTML
Added support for actions to navigate to new pages via data-actionUrl attribute (i.e. for "View Profile" action.)